requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
urllib3>=2.0.0